migrations changed, added timestamps

+++ src/main/resources/db/migration/V5__buildings.sql
@@ -2,6 +2,6 @@ CREATE TABLE buildings (id INT AUTO_INCREMENT PRIMARY KEY,
                         type VARCHAR(45),
                         level SMALLINT (4),
                         hp SMALLINT (4),
-                        started_at SMALLINT (10),
-                        finished_at SMALLINT (10) );
+                        started_at BIGINT (10),
+                        finished_at BIGINT (10) );
